test(codex): split live model discovery coverage into dedicated route tests
