+++ tests/unit/provider-models-route.test.ts
@@ -223,136 +223,6 @@ test("provider models route falls back after OpenAI-compatible endpoint probes a
   assert.ok(seenUrls.length >= 2);
 });
 
-test("provider models route discovers live Codex models and tags them for Responses", async () => {
-  const connection = await seedConnection("codex", {
-    authType: "oauth",
-    accessToken: "codex-access-token",
-    providerSpecificData: { workspaceId: "workspace-123" },
-  });
-  const seenRequests: Array<{
-    url: string;
-    authorization: string | null;
-    workspaceId: string | null;
-    originator: string | null;
-    userAgent: string | null;
-  }> = [];
-
-  globalThis.fetch = async (url, init) => {
-    const headers = new Headers(init?.headers as HeadersInit | undefined);
-    seenRequests.push({
-      url: String(url),
-      authorization: headers.get("authorization"),
-      workspaceId: headers.get("chatgpt-account-id"),
-      originator: headers.get("originator"),
-      userAgent: headers.get("user-agent"),
-    });
-    return Response.json({
-      models: [
-        {
-          slug: "codex-auto-review",
-          display_name: "Codex Auto Review",
-          visibility: "hide",
-          supported_in_api: true,
-        },
-        {
-          slug: "gpt-5.6",
-          display_name: "GPT 5.6",
-          visibility: "list",
-          supported_in_api: true,
-          max_input_tokens: 272000,
-          max_output_tokens: 128000,
-        },
-        { id: "", name: "missing-id" },
-      ],
-    });
-  };
-
-  const response = await callRoute(connection.id, "?refresh=true");
-  const body = (await response.json()) as any;
-
-  assert.equal(response.status, 200);
-  assert.equal(body.provider, "codex");
-  assert.equal(body.source, "api");
-  assert.deepEqual(seenRequests, [
-    {
-      url: "https://chatgpt.com/backend-api/codex/models?client_version=0.142.0",
-      authorization: "Bearer codex-access-token",
-      workspaceId: "workspace-123",
-      originator: "codex_cli_rs",
-      userAgent: "codex-cli/0.142.0 (Windows 10.0.26200; x64)",
-    },
-  ]);
-  const modelIds = new Set(body.models?.map((model) => model.id));
-  const liveModel = body.models?.find((model) => model.id === "gpt-5.6");
-  const syncedModels = await modelsDb.getSyncedAvailableModelsForConnection("codex", connection.id);
-  const syncedIds = new Set(syncedModels.map((model) => model.id));
-
-  assert.equal(liveModel?.name, "GPT 5.6");
-  assert.equal(liveModel?.inputTokenLimit, 272000);
-  assert.equal(liveModel?.outputTokenLimit, 128000);
-  assert.equal(liveModel?.apiFormat, "responses");
-  assert.deepEqual(liveModel?.supportedEndpoints, ["responses"]);
-  assert.ok(modelIds.has("gpt-5.5-low"));
-  assert.ok(modelIds.has("gpt-5.4-xhigh"));
-  assert.ok(syncedIds.has("gpt-5.6"));
-  assert.ok(syncedIds.has("gpt-5.5-low"));
-  assert.ok(syncedIds.has("gpt-5.4-xhigh"));
-});
-
-test("provider models route keeps Codex live discovery failures degraded", async () => {
-  const connection = await seedConnection("codex", {
-    authType: "oauth",
-    accessToken: "codex-access-token",
-  });
-
-  globalThis.fetch = async () => new Response("upstream unavailable", { status: 503 });
-
-  const response = await callRoute(connection.id, "?refresh=true");
-  const body = (await response.json()) as any;
-
-  assert.equal(response.status, 200);
-  assert.equal(body.provider, "codex");
-  assert.equal(body.source, "local_catalog");
-  assert.equal(body.intentional, undefined);
-  assert.equal(body.warning, "Codex live catalog unavailable — using local catalog");
-  assert.ok(body.models.some((model: any) => model.id === "gpt-5.5"));
-});
-
-test("provider models route returns cached Codex models when refresh discovery fails", async () => {
-  const connection = await seedConnection("codex", {
-    authType: "oauth",
-    accessToken: "codex-access-token",
-  });
-  await modelsDb.replaceSyncedAvailableModelsForConnection("codex", connection.id, [
-    {
-      id: "cached-live-codex",
-      name: "Cached Live Codex",
-      source: "imported",
-      apiFormat: "responses",
-      supportedEndpoints: ["responses"],
-    },
-  ]);
-
-  globalThis.fetch = async () => new Response("upstream unavailable", { status: 503 });
-
-  const response = await callRoute(connection.id, "?refresh=true");
-  const body = (await response.json()) as RouteBody;
-
-  assert.equal(response.status, 200);
-  assert.equal(body.provider, "codex");
-  assert.equal(body.source, "cache");
-  assert.equal(body.warning, "Codex live catalog unavailable — using cached catalog");
-  assert.deepEqual(body.models, [
-    {
-      id: "cached-live-codex",
-      name: "Cached Live Codex",
-      source: "imported",
-      apiFormat: "responses",
-      supportedEndpoints: ["responses"],
-    },
-  ]);
-});
-
 test("provider models route retries transient OpenAI-compatible probe failures before succeeding", async () => {
   const connection = await seedConnection("openai-compatible-retry", {
     apiKey: "sk-openai-compatible",
@@ -811,27 +681,6 @@ test("provider models route returns the updated local catalog for GitHub Copilot
     body.models.some((model) => model.id === "gpt-5.1"),
     false
   );
-});
-
-test("provider models route returns codex gpt-5.4 effort variants in the local catalog", async () => {
-  const connection = await seedConnection("codex", {
-    authType: "oauth",
-    apiKey: null,
-    accessToken: "codex-access",
-  });
-
-  const response = await callRoute(connection.id);
-  const body = (await response.json()) as any;
-  const modelIds = new Set((body.models || []).map((model: any) => model.id));
-
-  assert.equal(response.status, 200);
-  assert.equal(body.provider, "codex");
-  assert.equal(body.source, "local_catalog");
-  assert.ok(modelIds.has("gpt-5.4"));
-  assert.ok(modelIds.has("gpt-5.4-low"));
-  assert.ok(modelIds.has("gpt-5.4-medium"));
-  assert.ok(modelIds.has("gpt-5.4-high"));
-  assert.ok(modelIds.has("gpt-5.4-xhigh"));
 });
 
 test("provider models route returns the expanded local catalog for Kiro", async () => {
